Fix a display issue and a list issue with the mobile menu
In commit 25d6756 the Hanbot ID menu option was moved into an auth check.
Although it was moved into the auth check it was also moved into a roles
for loop. Therefore, if a user had more than 1 role it would show multiple
times. It is now in the auth check but outside the for loop.

Another issue was fixed where the mobile menu would show on desktop and
it would also show depending on the height of the mobile device being
used. The negative margin top being used to hide it was changed to a
larger negative number.