Change key file type and add error if Hanbot ID is not set
The response for user/keyfile/get_keyfile.php was changed so it had to be
accommodated otherwise the file would show undefined for it's content.

A 412 error was also added so if the user trys to get the keyfile without
setting their Hanbot ID first an error will be displayed.