Add information for payment types across the dashboard
When admins and developers are looking at their sales they will now be
able to see crypto payments among them.

The same goes for when looking at a user's profile. It will show the correct
information for crypto payments that the user has made.

Some small CSS fixes were done. This was mainly to fix padding issues that
were noticed.