Make the auth module and dev loader downloadable
The dev loader will now return a zip file so we needed to accommodate that
on the front end by making it downloadable. To keep the flow of the application
consistent I have also opted for the auth module to be downloadable in
case it also changes in the future.