Rework the system for free scripts
There wasn't a good system for free scripts or just simply getting rid
of a key length a seller didn't want offered anymore. A new system was
put into place where a lifetime price of 0 would be considered a free
script and to remove a key length they would need to input a price of
-1 for the respective time.

Frontend changes were also introduced to accomodate free scripts. They
will display on the UI that they are free and will no longer be allowed
to be added to a users cart.